pyproject.toml
[project]
name = "sistema-logistica"
version = "0.1.0"
description = ""
authors = [
    {name = "DeyvidAugusto",email = "deyvidaugusto100@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-filter (>=25.2,<26.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
